requirements.txt
accelerate==0.29.2
cryosparc-tools==4.4.1
easydict==1.9
flask==3.0.3
matplotlib==3.8.4
mrcfile==1.5.0
natsort==8.4.0
numba==0.59.1
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
psutil==5.9.0
pyFFTW==0.13.1
PyYAML==6.0.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.1
seaborn==0.13.2
tensorboard==2.16.2
timm==1.0.3
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.2
